Update API for roles (#611)
* remove `guard_name` from api and add id to transformer

* disallow update/ delete for root admin role via api

* disallow assigning root admin via api

* add api to remove user roles

* fix assignRoles & removeRoles

+++ app/Http/Controllers/Api/Application/Roles/RoleController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Api\Application\Roles;
 
+use App\Exceptions\PanelException;
 use Illuminate\Http\Response;
 use Illuminate\Http\JsonResponse;
 use App\Models\Role;
@@ -21,8 +22,8 @@ class RoleController extends ApplicationApiController
     public function index(GetRoleRequest $request): array
     {
         $roles = QueryBuilder::for(Role::query())
-            ->allowedFilters(['name'])
-            ->allowedSorts(['name'])
+            ->allowedFilters(['id', 'name'])
+            ->allowedSorts(['id', 'name'])
             ->paginate($request->query('per_page') ?? 10);
 
         return $this->fractal->collection($roles)
@@ -67,6 +68,10 @@ class RoleController extends ApplicationApiController
      */
     public function update(UpdateRoleRequest $request, Role $role): array
     {
+        if ($role->isRootAdmin()) {
+            throw new PanelException('Can\'t update root admin role!');
+        }
+
         $role->update($request->validated());
 
         return $this->fractal->item($role)
@@ -81,6 +86,10 @@ class RoleController extends ApplicationApiController
      */
     public function delete(DeleteRoleRequest $request, Role $role): Response
     {
+        if ($role->isRootAdmin()) {
+            throw new PanelException('Can\'t delete root admin role!');
+        }
+
         $role->delete();
 
         return $this->returnNoContent();
